backend/pyproject.toml
[project]
name = "guess-game"
version = "0.1.0"
description = "Guess game written on Django."
requires-python = ">=3.14"
dependencies = [
    "Django==5.2.11",
    "djangorestframework==3.16.1",
    "django-filter==25.2",
    "uvicorn==0.40.0",
    "psycopg2-binary==2.9.11",
    "drf-spectacular==0.27.2",
    "celery==5.6.2",
    "redis==7.1.1",
]